+++ SecretStoreBackup.psm1
@@ -1,11 +1,4 @@
 
-#region Main
-
 Get-ChildItem -path $PSScriptroot\functions\*.ps1 |
 ForEach-Object { . $_.Fullname}
-
-#endregion
-
-#define aliases
-