+++ Changelog.md
@@ -1,1 +1,7 @@
 # Changelog for SecretStoreBackup
+
+# 0.0.1
+
++ Initial files and module testing
++ Added help stubs
++ Added alias `xss` for `Export-SecretStore` and `iss` for `Import-SecretStore`